Fixed drawing transparent -itembackground colors by erasing to the tree -background color first if a transparent gradient is being drawn.
Fixed drawing gradient -itembackground colors in the whitespace region; gradients were only being
drawn in the visible part of the window, resulting in "squished" gradients when a column was partly
scrolled offscreen.

Distinguish between the area of a gradient brush and the area being painted so parts of a rectangle
can be painted with a larger gradient brush.

Remember the Tcl_Obj rep of the -fill and -outline marquee options.

Documented the -fill, -outline and -outlinewidth marquee options.