Implemented item-relative and column-relative gradient coordinates.
Replaced the "content" gradient coordinate type with the more generic "area"
type which takes the same argument as the [bbox] widget command.

Exit early when filling rectangles with gradients if the gradient brush is zero-
sized or the gradient has no color stops.

New utility function TreeArea_FromObj gets a TREE_AREA_xxx from an object.
Used by the [bbox] command as well as the "area" gradient coordinate type.

Replaced some seemingly-unneeded binary searches for RItems with direct
lookup, not sure what I was thinking when I wrote that old code.

Created a new flag for DItems to indicate whether that part of the item
contains valid display info, to fix a bug where I was checking rangeFirstD
which might have been set to NULL due to ranges being recalculated.

Added 2 DItem flags and 2 corresponding functions to set those flags which
tell the display code to redraw an item if the list is scrolled.  This is needed
for "area" gradient coordinates which aren't relative to the canvas.

Calculate gradient brush bounds properly when drawing the marquee.

Added a bunch of tests of the [gradient] command.

Demos:
- Added 2 new demos, gradients.tcl and gradients2.tcl.
- Used -canvaspadx, -canvaspady, -itemgapx and -itemgapy in more demos.
- Fixed off-by-1 error in imovie.tcl when editing the clip name caused by
  recent changes.
- Added gradient in mycomputer.tcl.