Tweaked display of column headers when available width is less than needed.
For a single line of text, the text plus ellipsis is no longer centered
in the available space.  This prevents the text from jumping around as
the column is resized.

A side effect of this is that the column layout can get smaller than the
smallest text plus elipsis "X...", whereas before the side=right sort
arrow would get pushed off the right of the column.

Clipping is used when drawing the text to prevent the sort arrow being
drawn overtop the text.