Fixed column header size changes not always being detected properly.
Various display functions would call Range_RedoIfNeeded to redo item layout
when the DINFO_REDO_RANGES flag was set.  However, some changes that would
cause DINFO_REDO_RANGES to be set were only handled in the Tree_Display
idle-time handler.

This caused problems when, for example, a column's width was changed then
the [item bbox] command was called before the next Tree_Display happened.
In this case the change to the width of the column was not seen and the
ranges weren't updated, resulting in [item bbox] giving outdated results.