Fixed badly-broken minimum-size calculation of styles.
On 2006-11-22 the "Changed Style_NeededSize..." commit broke the calculation
of minimum sizes of styles when one or more elements had the -squeeze layout
option.  After that change, only -squeeze=x -orient=horizontal minimum size
was calculated correctly (the only situation used in the demos).

Added a bunch of tests to ensure this stays fixed.