Maintain a list of DItems for headers just as for items.
Now individual column headers can be redrawn rather that redrawing the
entire header area.

Previously the list of DItems for headers was recreated for every call to
Tree_Display.  Now the DItems are updated, invalidated, and freed as needed.

Scrolling of headers is now as efficient as it is for items, only redrawing
the exposed edges of the headers.