New style layout option -center.
This option makes it possible to center elements while still having other
elements (or uneven padding) on one or both sides of the centered elements.

All the elements between the first and last ones with -center layout are
positioned as a group.  Elements outside of the centered group of elements
are positioned within the space available to them.