Revised the memory allocator for proper pointer alignment [Bug 3461526].
User 'matzek' was getting sporadic address-alignment crashes on HP-UX ia64
after the new header code was introduced.  I figured this was the result of
new structs introduced with the header code exposing a bug in the memory
allocator.  I revised the memory allocator based on the newer Tcl allocator
in TclAlloc.c.