Changes to the height of headers/borders preserves the scroll position.
The x/y scroll position has always been relative to the top-left corner of
the window, when it should really be relative to the top-left corner of the
content area (i.e., the area inside the borders and headers).

TreeCtrl.xOriginV2 and .yOriginV2 (version 2) now track the scroll position
relative to the top-left corner of the content area.  The result is that
changing the height (or visibility) of the headers or borders preserves the
scroll position.

In every previous version, showing the header would often result in the item
visible at the top of the content area becoming obscured by the header.  The
developer was often forced to add a call [$T yview moveto 0] as a workaround
after creating a treectrl and filling it with items.